# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Nature API
  version: 0.1.0
  description: |
    Nature 백엔드 API.

    ## 이 문서의 위치

    이 파일은 `//go:embed docs/openapi.yaml` 로 바이너리에 박혀 `GET /openapi.yaml` 로
    그대로 나가고, `GET /docs` 의 Swagger UI 가 이것을 읽는다. **파일이 없으면 빌드가 깨진다.**

    사용자 인증은 초대 전용 이메일·비밀번호 방식이다. 어드민 경로는 스텁이며, 요청·응답 스키마의
    정본은 `internal/admin` 의 핸들러다 — 여기에 상세를 지어 넣으면 구현과 어긋난다.
servers:
  - url: https://nature-api.redhead.kr
    description: 운영
  - url: http://localhost:8091
    description: 로컬 개발 (Firestore 에뮬레이터 연동)
security:
  - bearerAuth: []
paths:
...
  /admin/admins:
    get:
      tags: [admin]
      summary: 어드민 계정 목록
      description: 구현은 internal/admin 참고.
      operationId: listAdmins
      responses:
        '200':
          description: 목록
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      tags: [admin]
      summary: 어드민 계정 생성
      description: 구현은 internal/admin 참고.
      operationId: createAdmin
      responses:
        '200':
          description: 생성된 계정
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /admin/admins/{adminId}:
    parameters:
      - name: adminId
        in: path
        required: true
        description: 어드민 계정 문서 ID
        schema: { type: string, minLength: 1 }
    patch:
      tags: [admin]
      summary: 어드민 계정 수정
      description: 구현은 internal/admin 참고.
      operationId: updateAdmin
      responses:
        '200':
          description: 수정된 계정
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /admin/audit-logs:
    get:
      tags: [admin]
      summary: 감사 로그 조회
      description: 구현은 internal/admin 참고.
      operationId: listAuditLogs
      responses:
        '200':
          description: 감사 로그 목록
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        액세스 토큰(Bearer). 사용자 토큰과 어드민 토큰은 **서로 다른 시크릿으로 서명한다** —
        `ADMIN_JWT_SECRET` 이 `JWT_SECRET` 과 같으면 어드민 토큰이 사용자 미들웨어에서도
        파싱되므로 서버가 기동을 거부한다.
  responses:
...
    BadRequest:
      description: 요청 형식 또는 값 검증 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: VALIDATION_FAILED, message: 요청 값이 올바르지 않아요 }
    Unauthorized:
      description: 토큰이 없거나 유효하지 않음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: UNAUTHORIZED, message: 로그인이 필요해요 }
    InternalError:
      description: 서버 내부 오류
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: INTERNAL_ERROR, message: 서버 오류가 생겼어요 }
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: |
            기계 판독용 에러 코드. 클라이언트는 이 값으로 분기하고 `message` 는 그대로 표시한다.
            공통 코드는 `VALIDATION_FAILED` · `UNAUTHORIZED` · `INTERNAL_ERROR` 이며,
            사용자 인증은 `INVALID_CREDENTIALS` · `ACCOUNT_DISABLED`를 더한다.
            `TOO_MANY_ATTEMPTS`는 향후 시도 제한용 429 예약 코드다. 현재 제한은 구현하지 않았으며,
            구현 시 `details`에 남은 초를 담는다.
          example: VALIDATION_FAILED
        message:
          type: string
          description: 사용자에게 그대로 보여줄 수 있는 한국어 문구
          example: 요청 값이 올바르지 않아요
        details:
          type: object
          nullable: true
          additionalProperties: true
          description: 필드별 위반 내역 등 부가 정보
...